requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
openai>=1.30.0
ddgs>=0.1.0
duckduckgo-search>=8.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
